test(core): split pre-analysis tests into dedicated file (CodeRabbit #281)
Moves the new "EvolutionCycle pre-analysis wiring" describe block from
evolution-cycle.test.ts into a sibling evolution-cycle-preanalysis.test.ts
to keep both files under the 500-line repo budget (CLAUDE.md convention).
No test logic changes.

Also: silence biome lint/complexity/useLiteralKeys in errors.ts where
bracket notation is required for `noPropertyAccessFromIndexSignature`
(rule got escalated from info to error in biome 2.4.14 vs the 2.4.10
config schema). The existing comment already documents the intent;
this just adds the matching biome-ignore directive.
